Put cfg structs into separate file
